api/requirements.txt
asgiref==3.7.2
async-timeout==4.0.2
boto3==1.26.156
botocore==1.29.156
certifi==2023.5.7
Django==4.2.2
django-cors-headers==4.1.0
django-extensions==3.2.3
django-filter==23.2
django-redis==5.3.0
django-storages==1.13.2
djangorestframework==3.14.0
jmespath==1.0.1
psycopg2==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.5
s3transfer==0.6.1
sentry-sdk==1.25.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
urllib3==1.26.16
